requirements.txt
charm-tools
charmhelpers
charms.reactive
pre-commit
tox
